requirements.txt
# Flask核心框架
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0

# 数据库驱动
PyMySQL>=1.1.0
cryptography>=41.0.0

# 对象存储
minio>=7.1.0

# 图片处理
Pillow>=10.0.0

# 工具库
python-dotenv>=1.0.0
Werkzeug>=2.3.0

# 可选：输入验证（推荐）
# marshmallow>=3.20.0
# marshmallow-sqlalchemy>=0.29.0
